Check Leanpub Course

Check Leanpub Course

check_course(
  course_dir = ".",
  save_metrics = TRUE,
  timezone = "America/New_York",
  require_authorization = TRUE,
  use_book = FALSE,
  check_youtube_links = TRUE,
  ...
)

Arguments

course_dir
directory with course materials

save_metrics
Should an `rds` file be saved of the `data.frame`?

timezone
Timezone to be used?

require_authorization
Should you authorize for Google Slides and such to check the modified time?

use_book
Use Book.txt to keep only the manuscript files that were specified.

check_youtube_links
Should YouTube links be checked? If starting a course, set to FALSE

...
arguments to pass to didactr_auth

Value

A list of data frames of the checked course. Will have class course_check

Examples

root_path = tempfile()
course_name = "test this out2"
book_txt =  system.file("extdata", "Book.txt", package = "didactr")
sc = start_course(course_name, root_path, book_txt = book_txt)

#> ✔ Setting active project to '/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a160a6f25/test_this_out2'

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a160a6f25/test_this_out2/scripts directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a160a6f25/test_this_out2/manuscript/resources directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a160a6f25/test_this_out2/manuscript/resources/images directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a160a6f25/test_this_out2/manuscript/resources/videos directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a160a6f25/test_this_out2/metrics directory

#> ✔ Writing 'test_this_out2.Rproj'

#> ✔ Adding '.Rproj.user' to '.gitignore'

#> ✔ Adding 'manuscript/resources/videos' to '.gitignore'

#> ✔ Setting active project to '<no active project>'

course_dir = sc$course_dir
in_ci <- function() {
nzchar(Sys.getenv("CI"))
}
if (!in_ci()) {
object = check_course(sc$course_dir, require_authorization = FALSE)
}

#> Warning: `data_frame()` is deprecated as of tibble 1.1.0.
#> Please use `tibble()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Google Slides ID is missing from following lessons: 00_types, 01_reports, 02_presentations, 03_general, 04_technical, 05_blogpost, 06_meetings, 07_one_on_one, 08_notebook, 09_graphs