Check for expected file structure and create missing

Check for expected file structure and create missing

make_course(
  course_dir = ".",
  book_txt = NULL,
  verbose = TRUE,
  warn_book_exists = TRUE,
  create_directories = TRUE
)

check_structure(
  course_dir = ".",
  book_txt = NULL,
  verbose = TRUE,
  warn_book_exists = TRUE,
  create_directories = TRUE
)

Arguments

course_dir
directory with course materials

book_txt
A text file with names of markdown files and course names

verbose
print diagnostic messages

warn_book_exists
Should a warning be sent if Book.txt already exists.

create_directories
if the directories do not exist, create them.

Value

A list of the paths necessary for the course

Examples

course_dir = tempfile()
dir.create(course_dir)
make_course(course_dir)

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/scripts directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript/resources directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript/resources/images directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript/resources/videos directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/metrics directory

#> $course_dir
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1"
#> 
#> $scr_path
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/scripts"
#> 
#> $res_path
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript/resources"
#> 
#> $man_path
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript"
#> 
#> $img_path
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript/resources/images"
#> 
#> $vid_path
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript/resources/videos"
#> 
#> $met_path
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/metrics"
#> 
#> $book_txt
#> [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda81c9bcaa1/manuscript/Book.txt"
#> 
#> $course_name
#> [1] "filedda81c9bcaa1"
#> 
#> attr(,"class")
#> [1] "structure_check"