Starting a Course, Light wrapper for make_course

Starting a Course, Light wrapper for make_course

create_course(
  course_name,
  root_path = ".",
  book_txt = NULL,
  folder_id = NULL,
  verbose = TRUE,
  rstudio = TRUE,
  open = FALSE,
  git = FALSE,
  ...
)

Arguments

course_name
Name of the course. Any spaces will be set to underscores

root_path
The directory to put the course into.

book_txt
A text file with names of markdown files and course names

folder_id
Google ID for a folder of presentations

verbose
print diagnostic messages

rstudio
If TRUE, calls use_rstudio() to make the new package or project into an RStudio Project.

open
If TRUE, activates the new project: If RStudio desktop, the package is opened in a new session. Otherwise, the working directory and active project is changed.

git
If TRUE, calls use_git() to use Git inside the project.

...
additional argument to pass to create_lesson

Value

The output of make_course.

Examples

root_path = tempfile()
course_name = "test this out"
sc = create_course(course_name, root_path)

#> ✔ Setting active project to '/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda838141eb3/test_this_out'

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda838141eb3/test_this_out/scripts directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda838141eb3/test_this_out/manuscript/resources directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda838141eb3/test_this_out/manuscript/resources/images directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda838141eb3/test_this_out/manuscript/resources/videos directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda838141eb3/test_this_out/metrics directory

#> ✔ Writing 'test_this_out.Rproj'

#> ✔ Adding '.Rproj.user' to '.gitignore'

#> ✔ Adding 'manuscript/resources/videos' to '.gitignore'

#> ✔ Setting active project to '<no active project>'

root_path = tempfile()
course_name = "test this out2"
book_txt =  system.file("extdata", "Book.txt", package = "didactr")
sc = create_course(course_name, root_path, book_txt = book_txt)

#> ✔ Setting active project to '/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2'

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/scripts directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/resources directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/resources/images directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/resources/videos directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/metrics directory

#> ✔ Writing 'test_this_out2.Rproj'

#> ✔ Adding '.Rproj.user' to '.gitignore'

#> ✔ Adding 'manuscript/resources/videos' to '.gitignore'

#> ✔ Setting active project to '<no active project>'

readLines(sc$book_txt)

#>  [1] "00_types.md # Types of Communication"
#>  [2] "01_reports.md"                       
#>  [3] "02_presentations.md"                 
#>  [4] "03_general.md"                       
#>  [5] "04_technical.md ###Make a Techie"    
#>  [6] "05_blogpost.md   ###  Make a BLOG"   
#>  [7] "06_meetings.md"                      
#>  [8] "07_one_on_one.md"                    
#>  [9] "08_notebook.md"                      
#> [10] "09_graphs.md"                        

md_files = list.files(sc$man_path, full.names = TRUE)
md_files

#>  [1] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/00_types.md"        
#>  [2] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/01_reports.md"      
#>  [3] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/02_presentations.md"
#>  [4] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/03_general.md"      
#>  [5] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/04_technical.md"    
#>  [6] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/05_blogpost.md"     
#>  [7] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/06_meetings.md"     
#>  [8] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/07_one_on_one.md"   
#>  [9] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/08_notebook.md"     
#> [10] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/09_graphs.md"       
#> [11] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/Book.txt"           
#> [12] "/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpUTiJRp/filedda830eb9688/test_this_out2/manuscript/resources"          

readLines(md_files[1])

#>  [1] ""                                                                                                 
#>  [2] "# Types of Communication"                                                                         
#>  [3] ""                                                                                                 
#>  [4] "<!-- Google Slide ID -->"                                                                         
#>  [5] "<!-- SLIDEID -->"                                                                                 
#>  [6] ""                                                                                                 
#>  [7] "<!-- Include a slide PNG with Page_ID from this Slide Deck: -->"                                  
#>  [8] "<!-- ![](https://docs.google.com/presentation/d/SLIDEID/export/png?id=SLIDEID&pageid=PAGE_ID) -->"
#>  [9] "<!-- or use  `didactr::gs_slide_df(\"SLIDEID\")$png_markdown` -->"                                
#> [10] ""                                                                                                 
#> [11] "### Sections"                                                                                     
#> [12] ""                                                                                                 
#> [13] "SLIDESHERE"                                                                                       
#> [14] ""                                                                                                 
#> [15] "### Summary"                                                                                      
#> [16] ""                                                                                                 
#> [17] "### Additional Resources"                                                                         
#> [18] ""                                                                                                 
#> [19] "### Slides and Video"                                                                             
#> [20] ""                                                                                                 
#> [21] "<!-- ![Types of Communication](YouTube Link) -->"                                                 
#> [22] ""                                                                                                 
#> [23] "  - [Slides](Link to Slides)"                                                                     
#> [24] ""                                                                                                 
#> [25] "{quiz, id: quiz_00_types}"                                                                        
#> [26] ""                                                                                                 
#> [27] "### Types of Communication quiz"                                                                  
#> [28] ""                                                                                                 
#> [29] "{choose-answers: 4} "                                                                             
#> [30] "? What do you think?"                                                                             
#> [31] ""                                                                                                 
#> [32] ""                                                                                                 
#> [33] "C) The answer to this one"                                                                        
#> [34] "o) Not the answer"                                                                                
#> [35] "o) A bad answer"                                                                                  
#> [36] "o) Not an answer"                                                                                 
#> [37] ""                                                                                                 
#> [38] "{/quiz}"