Create a Lesson

Create a Lesson

create_lesson(
  lesson_name,
  course_dir = ".",
  verbose = TRUE,
  md_file = NULL,
  make_slide_deck = TRUE,
  slide_id = NULL,
  extract_code = FALSE,
  rmd = extract_code,
  open = FALSE,
  template_slide_id = "143gvqcynq_bl7iVd2G9yjumwJJkAy0S6CyNCsrJ2LgE",
  publish = TRUE,
  add_number = TRUE,
  use_google_drive = FALSE,
  ...
)

create_lessons_from_book(
  course_dir = ".",
  verbose = TRUE,
  make_slide_deck = FALSE,
  ...
)

Arguments

lesson_name
Name of the Lesson to create. This should be human-readable and will be the title of your lesson in your course.

course_dir
directory with course materials

verbose
print diagnostic messages

md_file
Output markdown file to create. If not specified, will take the lesson_name, sub out spaces, lower case it, and use that for the file name.

make_slide_deck
Create a slide deck on Google Slides if no link is provided.

slide_id
ID to slide deck on Google Slides. Caution, this will publish the slide deck to the web.

extract_code
If you have text with #rstats in the slide or #rstats in the Alt-text title, code will be included

rmd
Should an Rmd (Rmarkdown) be made versus a Markdown?

open
should the file be opened for editing?

template_slide_id
Slide ID for a template. This template will be copied to your new slide deck.

publish
should the slide be published so slides can be included in the markdown?

add_number
Add a prefix to the file name, such as `00_` for the first file

use_google_drive
passed to gs_slide_df to call gs_to_drive_pngs when putting in PNGs into the lesson.

...
arguments passed to check_didactr_auth

Value

A list of the created markdown manuscript file and script files. create_lessons_from_book will return a list of these lists, one for each lesson in the Book.txt file.

Examples

root_path = tempfile()
course_name = "test"
sc = start_course(course_name, root_path)

#> ✔ Setting active project to '/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a58d53551/test'

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a58d53551/test/scripts directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a58d53551/test/manuscript/resources directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a58d53551/test/manuscript/resources/images directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a58d53551/test/manuscript/resources/videos directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a58d53551/test/metrics directory

#> ✔ Writing 'test.Rproj'

#> ✔ Adding '.Rproj.user' to '.gitignore'

#> ✔ Adding 'manuscript/resources/videos' to '.gitignore'

#> ✔ Setting active project to '<no active project>'

verbose = TRUE
out =create_lesson(lesson_name = "how to Do Things",
course_dir = sc$course_dir,
make_slide_deck = FALSE,
open = FALSE)
dir(sc$man_path)

#> [1] "00_how_to_do_things.md" "Book.txt"               "resources"             

dir(sc$scr_path)

#> [1] "00_how_to_do_things_script.md"

readLines(sc$book_txt, warn = FALSE)

#> [1] "00_how_to_do_things.md"

readLines(out$md_file)

#>  [1] ""                                                                                                 
#>  [2] "# how to Do Things"                                                                               
#>  [3] ""                                                                                                 
#>  [4] "<!-- Google Slide ID -->"                                                                         
#>  [5] "<!-- SLIDEID -->"                                                                                 
#>  [6] ""                                                                                                 
#>  [7] "<!-- Include a slide PNG with Page_ID from this Slide Deck: -->"                                  
#>  [8] "<!-- ![](https://docs.google.com/presentation/d/SLIDEID/export/png?id=SLIDEID&pageid=PAGE_ID) -->"
#>  [9] "<!-- or use  `didactr::gs_slide_df(\"SLIDEID\")$png_markdown` -->"                                
#> [10] ""                                                                                                 
#> [11] "### Sections"                                                                                     
#> [12] ""                                                                                                 
#> [13] "SLIDESHERE"                                                                                       
#> [14] ""                                                                                                 
#> [15] "### Summary"                                                                                      
#> [16] ""                                                                                                 
#> [17] "### Additional Resources"                                                                         
#> [18] ""                                                                                                 
#> [19] "### Slides and Video"                                                                             
#> [20] ""                                                                                                 
#> [21] "<!-- ![how to Do Things](YouTube Link) -->"                                                       
#> [22] ""                                                                                                 
#> [23] "  - [Slides](Link to Slides)"                                                                     
#> [24] ""                                                                                                 
#> [25] "{quiz, id: quiz_00_how_to_do_things}"                                                             
#> [26] ""                                                                                                 
#> [27] "### how to Do Things quiz"                                                                        
#> [28] ""                                                                                                 
#> [29] "{choose-answers: 4} "                                                                             
#> [30] "? What do you think?"                                                                             
#> [31] ""                                                                                                 
#> [32] ""                                                                                                 
#> [33] "C) The answer to this one"                                                                        
#> [34] "o) Not the answer"                                                                                
#> [35] "o) A bad answer"                                                                                  
#> [36] "o) Not an answer"                                                                                 
#> [37] ""                                                                                                 
#> [38] "{/quiz}"                                                                                          

in_ci <- function() {
nzchar(Sys.getenv("CI"))
}
if (FALSE) {
# requires authorization
root_path = tempfile()
course_name = "test"
sc = start_course(course_name, root_path)
verbose = TRUE
out = create_lesson(lesson_name = "how to Do Things",
course_dir = sc$course_dir,
make_slide_deck = !in_ci(),
open = interactive())
dir(sc$man_path)
dir(sc$scr_path)
readLines(sc$book_txt, warn = FALSE)
readLines(out$md_file, warn = FALSE)
}

root_path = tempfile()
course_name = "test"
sc = start_course(course_name, root_path)

#> ✔ Setting active project to '/private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a159f8667/test'

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a159f8667/test/scripts directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a159f8667/test/manuscript/resources directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a159f8667/test/manuscript/resources/images directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a159f8667/test/manuscript/resources/videos directory

#> Creating missing /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmphFk73E/file239a159f8667/test/metrics directory

#> ✔ Writing 'test.Rproj'

#> ✔ Adding '.Rproj.user' to '.gitignore'

#> ✔ Adding 'manuscript/resources/videos' to '.gitignore'

#> ✔ Setting active project to '<no active project>'

verbose = TRUE
x =  system.file("extdata", "Book.txt", package = "didactr")
file.copy(x, sc$book_txt, overwrite = TRUE)

#> [1] TRUE

course_dir = sc$course_dir
res = create_lessons_from_book(course_dir = sc$course_dir, open = FALSE,
make_slide_deck = FALSE
)
dir(sc$man_path)

#>  [1] "00_types.md"         "01_reports.md"       "02_presentations.md"
#>  [4] "03_general.md"       "04_technical.md"     "05_blogpost.md"     
#>  [7] "06_meetings.md"      "07_one_on_one.md"    "08_notebook.md"     
#> [10] "09_graphs.md"        "Book.txt"            "resources"          

dir(sc$scr_path)

#>  [1] "00_types_script.md"         "01_reports_script.md"      
#>  [3] "02_presentations_script.md" "03_general_script.md"      
#>  [5] "04_technical_script.md"     "05_blogpost_script.md"     
#>  [7] "06_meetings_script.md"      "07_one_on_one_script.md"   
#>  [9] "08_notebook_script.md"      "09_graphs_script.md"       

readLines(sc$book_txt)

#>  [1] "00_types.md # Types of Communication"
#>  [2] "01_reports.md"                       
#>  [3] "02_presentations.md"                 
#>  [4] "03_general.md"                       
#>  [5] "04_technical.md ###Make a Techie"    
#>  [6] "05_blogpost.md   ###  Make a BLOG"   
#>  [7] "06_meetings.md"                      
#>  [8] "07_one_on_one.md"                    
#>  [9] "08_notebook.md"                      
#> [10] "09_graphs.md"                        

readLines(res[[3]]$md_file)

#>  [1] ""                                                                                                 
#>  [2] "# Presentations"                                                                                  
#>  [3] ""                                                                                                 
#>  [4] "<!-- Google Slide ID -->"                                                                         
#>  [5] "<!-- SLIDEID -->"                                                                                 
#>  [6] ""                                                                                                 
#>  [7] "<!-- Include a slide PNG with Page_ID from this Slide Deck: -->"                                  
#>  [8] "<!-- ![](https://docs.google.com/presentation/d/SLIDEID/export/png?id=SLIDEID&pageid=PAGE_ID) -->"
#>  [9] "<!-- or use  `didactr::gs_slide_df(\"SLIDEID\")$png_markdown` -->"                                
#> [10] ""                                                                                                 
#> [11] "### Sections"                                                                                     
#> [12] ""                                                                                                 
#> [13] "SLIDESHERE"                                                                                       
#> [14] ""                                                                                                 
#> [15] "### Summary"                                                                                      
#> [16] ""                                                                                                 
#> [17] "### Additional Resources"                                                                         
#> [18] ""                                                                                                 
#> [19] "### Slides and Video"                                                                             
#> [20] ""                                                                                                 
#> [21] "<!-- ![Presentations](YouTube Link) -->"                                                          
#> [22] ""                                                                                                 
#> [23] "  - [Slides](Link to Slides)"                                                                     
#> [24] ""                                                                                                 
#> [25] "{quiz, id: quiz_02_presentations}"                                                                
#> [26] ""                                                                                                 
#> [27] "### Presentations quiz"                                                                           
#> [28] ""                                                                                                 
#> [29] "{choose-answers: 4} "                                                                             
#> [30] "? What do you think?"                                                                             
#> [31] ""                                                                                                 
#> [32] ""                                                                                                 
#> [33] "C) The answer to this one"                                                                        
#> [34] "o) Not the answer"                                                                                
#> [35] "o) A bad answer"                                                                                  
#> [36] "o) Not an answer"                                                                                 
#> [37] ""                                                                                                 
#> [38] "{/quiz}"