Get Google Slide Elements IDs

Get Google Slide Elements IDs

gs_slide_df(id, extract_code = TRUE, use_google_drive = FALSE, ...)

notes_from_slide(id)

gs_notes_from_slide(id)

notes_from_slide_output(slides)

gs_get_slides(id)

gs_speaker_notes_id(id)

gs_replace_notes(id, notes)

Arguments

id
Slide id passed to get_slides_properties after passing through as_id

extract_code
If you have text with #rstats in the slide or #rstats in the Alt-text title, code will be included

use_google_drive
Use Google Drive to host the PNGs, calling gs_to_drive_pngs

...
additional arguments to pass to gs_to_drive_pngs

slides
an output of gs_slide_df(id)

notes
a character vector of notes to be added to the slides

Value

A data.frame of the identifiers and properties of the slides

Examples

if (FALSE) {
id = "1XoRj0pwaLI34XKZ7TljVeDHu-tbgGmXRmQa528JIwmw"
slides = gs_slide_df(id)
all_notes = notes_from_slide(id)
slides$slideProperties$notesPage$pageElements[[1]]
pe = slides$slideProperties$notesPage$pageElements
text = pe[[1]]$shape$text$textElements[[2]]$textRun$content
notes = sapply(pe,
function(r) {
res = sapply(r$shape$text$textElements, function(x) x$textRun$content)
res = unlist(res)
res = res[!is.na(res)]
if (is.null(res)) {
res = ""
}
paste(res, collapse = " ")
})
notes
}
if (FALSE) {
id = "1XoRj0pwaLI34XKZ7TljVeDHu-tbgGmXRmQa528JIwmw"
slides = gs_get_slides(id)
tfile = tempfile(fileext = ".txt")
writeLines(slides$content, con = tfile)
ss = strsplit(slides$content, split = "\n")[[1]]
ind = grep("uthinkk", ss)
ss[(ind -30):(ind + 5)]

slides$parsed$slides$slideProperties$notesPage
slides$parsed$slides$slideProperties$notesPage$pageElements

}
if (FALSE) {
id = "1XoRj0pwaLI34XKZ7TljVeDHu-tbgGmXRmQa528JIwmw"
parsed = gs_get_slides(id)$parsed
parsed$slides$slideProperties$notesPage$notesProperties$speakerNotesObjectId

notes_id = gs_speaker_notes_id(id)

}
if (FALSE) {

id = "1XoRj0pwaLI34XKZ7TljVeDHu-tbgGmXRmQa528JIwmw"
shape_ids = gs_speaker_notes_id(id)
notes = sample(letters, size = length(shape_ids))
res = gs_replace_notes(id, notes)
curr_notes = notes_from_slide(id)
all(curr_notes == notes)
}