Translate Course to Another Language

Translate Course to Another Language

translate_course(
  course_dir,
  target = "es",
  translate_slides = FALSE,
  detect = TRUE,
  verbose = TRUE,
  use_book = TRUE,
  overwrite = FALSE,
  sleep_time = 0,
  trash_same_gs_name = FALSE,
  ...
)

Arguments

course_dir
directory with course materials

target
language to translate to, see gl_translate

translate_slides
Should slides be translated as well? If so, copy_and_translate_slide will be done with the Google slide ID and replaced throughout the manuscript.

detect
Should gl_detect_file be run to check that the language is not currently the target? Must be google Language authorized using gl_auth

verbose
Print diagnostic messages

use_book
Use Book.txt to keep only the manuscript files that were specified, passed to check_course

overwrite
Should files be overwritten? If not, they are put in a temporary directory and translated.

sleep_time
time in seconds to sleep before running next call

trash_same_gs_name
passed to copy_and_translate_slide Should other Google slide decks with the same name be trashed before copying? If not, can fill up your drive.

...
additional arguments to pass to translate_manuscript and translate_script

Value

A result from check_course

Examples

if (FALSE) {
course_dir = "~/Dropbox/Projects/CDS/cds_dataanalysis"

didactr::check_didactr_auth()

res = translate_course(course_dir)

}